Move download count logic into model and view modules
