+++ src/models/mod.rs
@@ -2,7 +2,7 @@ pub use self::badge::{Badge, MaintenanceStatus};
 pub use self::category::{Category, CrateCategory, NewCategory};
 pub use self::crate_owner_invitation::{CrateOwnerInvitation, NewCrateOwnerInvitation};
 pub use self::dependency::{Dependency, DependencyKind, ReverseDependency};
-pub use download::VersionDownload;
+pub use self::download::VersionDownload;
 pub use self::follow::Follow;
 pub use self::keyword::{CrateKeyword, Keyword};
 pub use self::krate::{Crate, CrateDownload, NewCrate};
@@ -19,6 +19,7 @@ mod badge;
 mod category;
 mod crate_owner_invitation;
 pub mod dependency;
+mod download;
 mod follow;
 mod keyword;
 pub mod krate;
